configuration set from Azure App Configuration
