--- SchedulerJobs/SchedulerJobs.Common/Configuration/ConfigLoader.cs
@@ -1,16 +1,0 @@
-﻿using Microsoft.Extensions.Configuration;
-
-namespace SchedulerJobs.Common.Configuration
-{
-    public class ConfigLoader
-    {
-        public readonly IConfiguration Configuration;
-        public ConfigLoader()
-        {
-            var configRootBuilder = new ConfigurationBuilder()
-                   .AddJsonFile("local.settings.json", optional: true, reloadOnChange: true)
-                   .AddEnvironmentVariables();
-            Configuration = configRootBuilder.Build();
-        }
-    }
-}
